the program execution page
the program execution page, i.e. after clicking start the program displays the steps on the page in time. It needs a pause function and two modes: hand-operated mode and hand-free mode. In hand-operated mode, the user needs to click from one step to the next; in hand-free mode, the prompt is displayed for a certain period of time (e.g., 30 seconds) and then automatically jumps to the next step. When finished, the pop-up window jumps back to the Home page.